Unbreak milestone creation over Conduit API (project.edit).
This addresses a problem reported on the phabricator community web site[0]
and on the FreeBSD ports mailing list[1].

Creation would fail reporting "Call to a member function getPHID() on null".

The underlying issue was the code introduced in T13462[2][3], which aims to
set the correct members/policies when creating a milestone. This code
relies on the new object to be already initialized with a parent project.
When using the user interface, this is done in
PhabricatorProjectEditController[4]. When using the Conduit API, this
code path isn't used though, hence the error message when trying to use the
null-parent.

This fix catches the problem while creating a copy that is adjusted for
policy checks - this seemed the best solution, as it doesn't mess with
anything else and the correction is close to where the problem happens.

Using other mechanisms like overloading _applyInitialEffects_ won't work, as
they happen too late in applyTransactions.

[0]https://discourse.phabricator-community.org/t/call-to-a-member-function-getphid-on-a-non-object-while-creating-milestone-using-conduit/3370/4
[1]https://lists.freebsd.org/pipermail/freebsd-ports/2020-June/118730.html
[2]https://secure.phabricator.com/T13462
[3]https://github.com/phacility/phabricator/commit/d58eddcf0ad4bec4762632a351f6b604f0a6afb4
[4]https://github.com/phacility/phabricator/blob/master/src/applications/project/controller/PhabricatorProjectEditController.php#L84